Bug#29959567 WL#12680: NDB_CONFIG.TEST FAILS ON SOME PLATFORMS
Problem
=======

In the new binary configuration file format, version 2, a default section
is provided for each type of sections.

As default section one of the non default sections was choosen and node id
parameters excluded.

For optional configuration parameteres without any default value it could
happen that if one section did not have that parameter set while another
section had that parameter set, and the latter section was choosen as basis
for default section, also the former section could get the same parameter
setting as the latter.

Solution
========

When creating the default section, only include configuration parameters
that are set in all sections of same type.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>