Bug #28208285 COLLECTIONS REPORTED AS BEING TABLES WITH NO_BACKSLASH_ESCAPES
Description:
When running the MySQL server using the --sql-mode=NO_BACKSLASH_ESCAPES
option, the X Plugin seems to be reporting collections as being tables.

As a user, running the mysqlxtest script below, I expect the
Mysqlx.Resultset.Row message being sent by the plugin to report the "bar"
database object as being a "COLLECTION" and not a "TABLE".

RB: 20873
Reviewed-by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>
Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>

Change-Id: If6ea149da3e741ed3119e867771841079b7a60c1