Bug #30209596 GCS PROTOCOL CHANGE CRASHES IF SOME MEMBER DOESN'T SUPPORT DESIRED NEW PROTOCOL
Problem
========================================================================
MySQL 8.0.16 introduced the notion of the group's communication
protocol. Since MySQL 8.0.16 and as of the time of writing
(MySQL 8.0.18), we have not created a new communication protocol.

While doing an exploratory prototype for a worklog over MySQL 8.0.18, I
created a new communication protocol. Subsequent manual cross-version
testing uncovered this bug.

In a nutshell, if one creates a group consisting of some members that
support the new communication protocol (8.0.18) and some members that do
not support the new communication protocol (e.g., 8.0.17), and one tries
to change the protocol to 8.0.18, the members will crash.

Analysis
========================================================================
The issue is that even though GCS correctly identifies that the desired
communication protocol (8.0.18) is not supported by some members and
outputs the appropriate message to the error log, the protocol change
procedure is not actually aborted.

Solution
========================================================================
Properly abort the protocol change procedure, i.e. release the tagged
lock and notify any waiting threads.

Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed-by: Anibal Pinto <anibal.pinto@oracle.com>
RB: 22864