Bug #30184265 MYSQL-8.0: NDB RESTORE IS FAILING WITH ERROR ""RESTORE: FAILED TO LOAD CONTENT".
From: Frazer Clement <frazer.clement@oracle.com>
Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>

Improve ndb_restore to print the Ndb error number and message when
it is unable to load a table descriptor from a backup .ctl file.

This can happen for example when a backup includes a table using a
character set which is unknown on the version of ndb_restore being
used to restore it.

Showing a specific error message helps users understand the source
of the problem.

Change-Id: I4e5e066354f01dbc3b290bb70bc823818f63f531