Bug#30131161: SYNTAX ERROR WITH || WHEN USING PIPES_AS_CONCAT AFTER UPGRADING TO 8.0.17
Problem:
--------
The issue was an unexpected side effect of the fix of bug29305022 "THE GRAMMAR
RULE OPT_ESCAPE CAUSES 6 SHIFT/REDUCE CONFLICTS".
That fix has affected the precedence of the `||` string concatenation operator
in one unusual particular case: when used inside expressions of 2nd or 3rd
parameter of the `LIKE` operator:

  ... LIKE substr1 || substr2 ...

or

  ... LIKE pattern ESCAPE substr1 || substr2 ...

so, those cases have started to fail with a syntax error.

The issue caused a syntax error instead of e.g. wrong data, since the
parser reduced the query at the right of `||` to <simple_statement> in
the examples above, i.e.:

  <simple_statement> || substr2

and it failed, since <simple_statement> is not acceptable as the
left hand side of `||`.

Analysis:
--------
The `||` operator is very special: it can has different operator
precedences and different semantics depending on the value of
the PIPES_AS_CONCAT SQL mode flag:

* When the PIPES_AS_CONCAT is disabled (by default):

  - semantics of `||` is represented by the OR2_SYM token.

  - `||` has the same operator precedence as `OR` has.

* When PIPES_AS_CONCAT is enabled:

  - semantics of `||` is represented by the OR_OR_SYM token.

  - `||` is documented as having "a precedence between ^ and
    the unary operators".

OTOH, most likely by a mistake, the historical definition of the
OR_OR_SYM (or OR_OR_CONCAT) token in the parser grammar was forced to
have the same precedence as the `OR` operator from the very beginning:

  ... lowest priorities ...
  %left   SET_VAR
  %left   OR_OR_SYM OR_SYM OR2_SYM
  %left   XOR
  ...
  %left   EQ ... LIKE ...
  ...
  %left   '^'
  %left   NEG '~'
  %right  NOT_SYM NOT2_SYM
  ... highest priorities ...

That precedence declaration of OR_OR_SYM didn't have any effect because
of shift/reduce conflicts in the depth of <expr>.

After the fix of bug29305022, the precedence of OR_OR_SYM
(that was declared equal to the precedence of the operator `OR` by mistake)
has been taken into account:

  1277 predicate: bit_expr not LIKE simple_expr .  [... OR_OR_SYM, ...
  ...
  Conflict between rule 1277 and token OR_OR_SYM resolved as reduce (OR_OR_SYM < LIKE).

Thus, the parser has started to recognize

  ... LIKE 'foo' || 'bar'

similar to

  (... LIKE 'foo') || 'bar'

instead of

  ... LIKE ('foo' || 'bar')

That caused the current bug.

Fix:
----
The current patch has fixed the precedence of `||` according to the
documentation:

  ... lowest priorities ...
  %left   SET_VAR
  %left   OR_SYM OR2_SYM
  %left   XOR
  ...
  %left   EQ ... LIKE ...
  ...
  %left   '^'
  %left   OR_OR_SYM
  %left   NEG '~'
  %right  NOT_SYM NOT2_SYM
  ... highest priorities ...

(Note OR_OR_SYM after `^` and before `NEG`.)

That has resulted in a correct resolving (shifting `||` instead of
reducing <simple_expr>) as documented:

  1277 predicate: bit_expr not LIKE simple_expr .  [ACCOUNT_SYM, ...
  ...
  Conflict between rule 1277 and token OR_OR_SYM resolved as shift (LIKE < OR_OR_SYM).

Change-Id: I1ac3b3cfb85a14775829b04ec5d0d6694af390ab