Bug #29627690	NOT ABLE TO IMPORT PARTITIONED TABLESPACE FROM VERSIONS 8.0.14-16
PROBLEM
-------
This bug is a regression after the fix for
<Bug26925260 ERROR 1030 (HY000): GOT ERROR 44 - 'INNODB ERROR' FROM STORAGE ENGINE>

During import we first try to open the cfg file corresponding to the
tablename. To search for a cfg file,we first try to build the path
to cfg with the data directory and the tablename ,but because of the
above bug the partition table ibd file as well as the cfg file are built
in lower cases where as in version <=8.0.13 we build it in upper cases.
This is the reason we are unable to get the path to the cfg and ibd file

FIX
---

If we don't find the cfg/cfp file ,we search it after converting table
name to lower cases and if found we rename the cfg/cfp and ibd
file according to previous convention.

Approved by: Mayank Prasad <mayank.prasad@oracle.com>