Bug#30220974: CHANGE UNIT TESTS TO USE A PROPER BUFFER FOR FIELD NULL VALUES [noclose]
Currently, our Field unit tests often use separate buffers for storing
values and null. This is different from how Fields work in the server,
where (almost always) the same buffer is used. This difference makes
refactoring Field a bit more difficult.

This patch changes unit tests so that the same buffer is used for
values and null, so that the code becomes more similar to the server.

Change-Id: I035eaadc290a5ca69eb4cd7d60c8c2a4920e4f50