Bug#30227756: JSON.JSON_PERFSCHEMA FAILS ON WINDOWS
The json.json_perfschema test fails on Windows if it runs after the
json.array_index_debug test, indicating that some allocated data has
not been freed.

The reason is that some Field objects are not destructed. These are
the Field objects that live in the TABLE_SHARE object and Field
objects used by add_functional_index_to_create_list() to set up a
generated column to back a functional index.

These objects are usually empty and need no destruction. The exception
is Field_typed_array, which contains a std::vector (indirectly, via a
Json_array member). The vector is empty, and on most platforms this
means that it has not allocated any memory on the heap. On MSVC,
however, the construction of an empty std::vector instance seems to
allocate some memory on the heap. Since the vector is never destroyed,
this memory is leaked.

This patch moves the Json_array member of Field_typed_array into the
Item_func_array_cast class. Since Item objects are always destructed
when they reach their end of life, this ensures that the vector is
destructed.

Change-Id: I3454f664faa5381b2385439b427be4ad254f45fb