BUG#30032129 ADD TEST THAT ERRORS IF APPLIER USER TRIES TO `SUID` TO A SYSTEM USER ACCOUNT
Description
-----------
The configured applier privilege context, although it may have `SET_USER_ID`
shouldn't be able to SUID to a user that is a system user if it's not privileged
with `SYSTEM_USER`.

Fix
---
A test script must be added to guard this use-case and expose regressions around
this topic.

Changed the `rpl_init.inc` infra-structure in order to allow to configure
additional privileges for the `PRIVILEGE_CHECKS_USER` user.

Changed the `rpl_init.inc` infra-structure in order to add `INSERT` on
`mtr.test_suppressions` privilege to `PRIVILEGE_CHECKS_USER` and allow the call
to `mtr.add_suppression` to be properly replicated.

Reviewed-by: Pedro Ribeiro <pedro.a.ribeiro@oracle.com>
Reviewed-by: Abhinav Agarwal <abhinav.ab.agarwal@oracle.com>
RB: 22861