Bug#30240381 Test ndb.apply_stored_grants fails occasionally
The test can fail with a "node unsubscribed" warning in the results.
The timing of the warning is not predictable, so the fix is to
disable warnings for code 1296 (ER_GET_ERRMSG) for the entire test.

The test can also fail due to a race condition if check_testcase
checks the contents of the mysql.user table on mysqld2.1 before
the cleanup of users has been applied there. Connect once more
to mysqld2 at the end of the test. The connection will have to
wait for a read lock on the ACL cache, hopefully ensuring that
the cleanup task, which should have already taken an exclusive
lock, will complete prior to check_testcase.

Approved-By: lakshmi.narayanan.sreethar@oracle.com
Change-Id: I8a7a56b5891b8b995f29fd454240b4f348c52485