Bug #29846292	ROLLBACK OF BIG TRANSACTION DUE TO CONFLICT REMAINS IN HUNG STATE
ROLLBACK of an INSERT of a LOB larger than a buffer pool caused a deadlock.
This changeset applies solution used previously for lob::z_rollback() as fix for
similar Bug #27321176 INNODB: SEMAPHORE WAIT . WAIT IN LOB ROLLBACK,
to the lob::rollback() function.
As the original fix had a small bug - leading to a assertion failure if a crash happened when
the actual lob size was smaller than DICT_MAX_FIELD_LEN_BY_FORMAT_FLAG - the approach was slightly modified,
so we now first mark the LOB as partially deleted by setting its length to zero,
before proceeding to delete its parts.

Reviewed-by: Rahul Agarkar <rahul.agarkar@oracle.com>
RB:22643