Bug #30120858 NDB : GSL RETRIES SHOULD REFRESH SCHEMA DEFINITION
The Ndb_table_guard class uses RAII to guard the scope of use
of table definitions from the NdbApi 'Global' dictionary by
providing wrapped access to an NdbApi table object pointer.

The current design of the Ndb_table_guard assumes that the
underlying object pointer is only determined once, at initialisation,
and afterwards, the previously retrieved pointer is returned
from a cached reference.

This means that a single Ndb_table_guard cannot be used in situations
where the pointer retrieval might need to be retried.

In this case, a separate Ndb_table_guard must be used for each
attempt.

The Global Schema Lock (GSL) code was using a single Ndb_table_guard object
to retry attempts to get a table object reference, and this can never
succeed if it fails on first attempt.

This resulted in infinite waits to obtain the GSL, causing the
Binlog Injector to get stuck so that the MySQL Server considers
Ndb tables to be read only.

Fix proposed here uses a fresh instance of Ndb_table_guard for
each retry.

Testcase ndb_restart_debug is created to test this fix.

Approved by : Magnus Blaudd <magnus.blaudd@oracle.com>