Bug #30254797: REPLACE VECTOR<QEP_TAB *> WITH BITMAPS
There are several places in the translation of pre-iterator executor code (in
particular, to hash joins, but also in a few other places) where we should
juse use a bitmap. Do so, and also add a convenience class for looping over
such bitmaps using range-based for loops.

This saves ~1.3 kB of generated code in an optimized build.

Change-Id: I7fa2a8982f51d2dcfc6fe1293156a644fb5ba0ef