BUG#30254755 STRENGTHENS CPCD PROCESS LIFECYCLE
Prevents CPCD::undefineProcess to implicitly stop a process.
Call both CPCD::stopProcess and CPCD::undefineProcess when stopping a
session.

By removing implicit stop from undefineProcess we're ensuring that each
CPCD API call (defineProcess, undefineProcess, startProcess, and
stopProcess) is responsible for a single state transation, hence
simplifying/strenghthen process' lifecycle.
With this chage we'll be rejecting any attempt to undefine a process
that has not been previously stopped.

The above change, in particular the rejection to undefine a non-stopped
process, requires both CPCD::stopProcess() and CPCD::undefineProcess()
calls in CPCDAPISession::stopSession().
Calling CPCD::stopProcess() to an already stopped process will generate
an error (which we'll ignore) and we'll be logging that error.
That is expected as this provides visibility that a session didn't
properly clean processes it created and that requires further
investigation.

Change-Id: Ie074e102da2bf64006f63c8b87abacd785a57644