Bug#30040815: SERVER STARTUP ON WIN FAILS IF DATA_HOME_DIR AND DATADIR USES DIFF LETTER CASES
Post-fix for two build problems.

Doxygen build broken by a missing @param

Werror build broken by an unnecessary attempt to recover from a failed call
to my_realpath().  The change used strncpy which should not be used since it does not
guarantee a null termination.  Reverted this part of the change since it was not
needed for the fix.

RB#22834