Bug#30040815 - Post-fix
This patch revision a post-fix change which;
1. fixes one testcase,  meb.add_drop_add_replay started to fail because it needed to suppress a new message
2. fixes some recent MSVC compiler warnings,
3. improves Fil_path::get_real_path().

The patch pushed to mysql-8.0 in ff9ab65d716 contained a change to get_real_path() that causes a build failure on the werror build in mysql-8.0.  That change was reverted in [3076a1f6870] because I did not think it was needed.  But it IS needed for Bug#30040815 in MacOS and shows up as a bunch of test failures in the meb suite.

my_realpath() will fail on MacOS and Linux when a folder does not yet exist.  When it fails, Fil_path::get_real_path() just returns the path provided.  But with this bug fix, the path provided needs to be made lower case on case insensitive file systems like on MacOS.  The MEB suite seems to test this scenario.

So the change to Fil_path::get_real_path() is put back in with a safer method of copying the provided path into a local buffer so it can be made lower case. It uses memcpy and setting
the null terminator instead of using strncpy

Approved by Marcin in RB#22834