Bug#30261942 IMPROVE SUPPORT FOR "SYSTEM" OPENSSL ON WINDOWS
"system" OpenSSL on Windows is supported if an external OpenSSL
package is installed.  The obvious place to install it is
  C:/Program Files/OpenSSL-Win64
This however, currently does not work, the whitespace creates problems.

The fix is to copy the .lib files to the build directory,
and link with the copies.

Also: handle WITH_SSL=yes, and convert it to WITH_SSL=system

Change-Id: I1db7c30ad4d83eef0884687a5cabbc276ecab770