Bug#30262505: REMOVE STORE_KEY::ERR
The 'err' member of the store_key class was in some cases used to
temporarily store null values for Fields. This made it harder to
change how Field state is represented.

This patch refactors store_key to replace usage of 'err' with
Field::set_tmp_nullable() and Field::is_tmp_null(), which is
the way we typically store null values in non-nullable Fields.

Change-Id: I794394c22e1f89d455d2360f5e37132da4693b4f