Bug#30272247: CHANGE ITEM_FUNC_SP TO USE A PROPER RECORD BUFFER
Item_func_sp uses a Field to store the result of a stored function.
This Field did not use a proper record buffer for storage.
The value itself was stored in one buffer and the null value was
stored in a separate bool.

This makes it more difficult to change how Fields store data.
This patch refactors Item_func_sp so that its Field now has
a proper record buffer.

Change-Id: I5409387d03e87477aca062b4c0d3205641b360e8