Bug#29305977 TEST NDB.NDBINFO_UPGRADE TEST FAILS
When mysqld is run with the --upgrade=force option, it reports:
 [Warning]  Table 'mysql.ndb_apply_status' requires repair.
 [ERROR]    Table 'mysql.ndb_apply_status' repair failed.

The --upgrade=force option causes a bootstrap system thread to run
a "CHECK TABLE FOR UPGRADE" command, but ha_ndbcluster::open() refuses
to open the table before schema synchronization has completed, which
eventually leads to the "requires repair" and "repair failed" messages.

For this fix, ha_ndbcluster::open() is refactored to fit about one
screenful of code. The refactorings are:
 - Extract method open_table_set_key_fields().
 - If open_table_set_key_fields() fails, return HA_ERR_OUT_OF_MEM
   rather than 1.
 - Extract methods release_key_fields() and release_ndb_share() from
   local_close(), and then remove method local_close().
 - m_table_map is acquired in get_metadata(), so its release is moved
   from local_close() into release_metadata().
 - Extract method open_table_before_schema_sync().

The actual logic for fixing the bug is in open_table_before_schema_sync(),
testing (thd->system_thread == SYSTEM_THREAD_SERVER_UPGRADE).

Change-Id: Ic2e70fb7b4208184408cc5681347dc6524124941