Bug #30273183 	CRASH ON RELATIVE INNODB_DIRECTORIES
The patch for Bug#30040815 added an assert that programmatically,
the paths provided to Fil_path::is_ancestor() should be absolute paths.
But this bug exposed a condition in which Fil_path::get_real_path() did not
make the conversion correctly.

That condition happens when --innodb-directories refers to a path that does
not yet exist on non-Windows platforms.  Windows can make a realpath from a
path that does not exist, but other platforms fail.

So what happens is that after startup, the list of scan directories contains
a directory that has not yet been converted to an absolute path and made to
lower case on case insensitive file systems.  This causes the assert added in
Bug#30040815 to fail.

The solution implemented here is to make Fil_path::abs_path() call get_real_path()
if m_abs_path was not set correctly when the Fil_path object was constructed.
The Fil_path object is handled as a const object so m_abs_path is not reset.
This is OK since it is rare that m_abs_path will be null.  That will only happen on
non-Windows platforms when the path did not exist at startup.

Approved by Marcin and Ingo in RB#22940