Bug #29846292 ROLLBACK OF BIG TRANSACTION DUE TO CONFLICT REMAINS IN HUNG STATE
Post-push fix for "partial update".

The fix for Bug #29846292, in order to indicate that a blob version is being
rolled back, as one of the first steps, changes the length of the blob declared
in the blob reference to 0. This was done, so that if a crash happens during the
rollback, and some pages of the blob are removed, while others are still not,
then during the subsequent recovery we can identify this situation, by seeing
that there is a blob reference, but it has zero length, and thus we should refrain
from reading the (partially deleted and "inconsistent") blob version and instead
finish the clean up by removing any left-overs.

However, there was an assert in the lob::rollback(), which compared the
length stored in the blob reference to the actual length of the blob, and this
assert now fails.
The assert is only verified if the lob was modified via "partial update",
as opposed to:
- "inserted", during INSERT
- "fully modified", when the UPDATE changes blob so much that we create a new one
- modified with "small change optimization", in which case we store the difference
  in the undo record itself instead of creating new version via partial update

Unfortunately, none of the existing tests covered the "partial update" scenario.
In particular the innodb.blob_partial_update, despite its name, involved
modifications shorter than LOB_SMALL_CHANGE_THRESHOLD and thus triggered
the "small change optimization".

This changeset:
- Fixes the bug in which ref.length for the assert was read after being set to 0.
  Now, the old value of ref.length from before modification will be used in the assert.
- Also, makes sure, that if ref.length before modification actually happend to be 0,
  it is correctly interpreted as a symptom of a crash during previous rollback(),
  and the cleanup continues wihout an assertion failure.
- Introduces innodb.blob_partial_update_2 which executes scenarios similar to those in
  in innodb.blob_partial_update, but makes sure to invoke "partial update" and not the
  "small change optimization". It also checks explicitly what happens if we
  ROLLBACK each query.
- Introduces innodb.lob_rollback_crash_4 testing a crash happening during rollback of
  a "partial update".
- Introduces innodb.lob_rollback_crash_5 testing a crash happening during "full update".
- Introduces innodb.zlob_rollback_crash_5 testing a crash happening during rollback of
  a "partial update" of a compressed lob.
- Adds information about iteration number during the main loop of rollback when the
  lob_rollback_print_index_size logging is enabled.
- Changes the way crash_middle_lob_rollback determines the moment to crash. Old code
  looked at the length of the index, new code will check iterations count. This way
  partial updates, which do not modify length of the index, can be tested for a crash
  "in the middle" of the loop.

Reviewed-by: Rahul Agarkar <rahul.agarkar@oracle.com>
RB:22924