BUG#29969769 - KILL QUERY KILLS THE NEXT QUERY
KILL QUERY may possibly KILL the next query. This happens
when a KILL query comes in after cleanup in mysql_execute_command
before the m_server_idle flag is set. The next query may start with
the killed flag and the query shall get killed.
The fix is to reset the killed flag before the execution of the query
begins. The test thread_pool_kill_coverage is modified to account
for the behavior KILL QUERY always terminates the statement that
is in execution.

Change-Id: I19e0facdbceb8465f201ca34e462e56ed01063bd