Bug #30139794 STOP GR HANGS ON A MEMBER WHOSE XCOM PORT IS BLOCKED.
Problem
========================================================================
STOP GROUP_REPLICATION hangs on a member whose XCom port is blocked. By
blocked we mean that Linux's iptables was configured to drop all
incoming or outgoing packets to the port, e.g.:

  # iptables -A INPUT  -p udp --dport <port> -j DROP
  # iptables -A INPUT  -p tcp --dport <port> -j DROP
  # iptables -A OUTPUT -p udp --dport <port> -j DROP
  # iptables -A OUTPUT -p tcp --dport <port> -j DROP

Analysis
========================================================================
GCS sends XCom a termination command via XCom's input queue to force
XCom to exit. However, the input queue's signalling connection uses the
XCom's port, which means that the signal will never reach XCom if the
port is blocked.
If XCom's local_server task is waiting for the signal on the signalling
connection, it will not wake up, hence some user/GCS thread hangs
waiting for the XCom thread to terminate.

Solution
========================================================================
Rely solely on the get_should_exit callback to forcefully terminate
XCom, since it is already the fall back which is supposed to always
work.
Terminating XCom via the callback or the termination command is
equivalent.

RB: 22905
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
Reviewed-by: Aníbal Pinto <anibal.pinto@oracle.com>