Bug#30292605: ASSERTION WHEN READING A HIDDEN ARRAY COLUMN
When using the debug option "show_hidden_columns" to expose the hidden
generated columns that back functional indexes, an assert failure is
triggered while reading the columns that back the multi-value indexes.

The problem is that these columns are attempted sent as the array
element type instead of as an array data type. For example, an array
of integers is attempted sent as an integer instead of as an array.

Fixed by sending the columns as JSON arrays.

Change-Id: I960542e6daf11e80feb4b5bcb1b8bb7fcb7ff02c