Bug#27664092 - UNEXPECTED GCP STOP WHEN OTHER NODE SHUTS DOWN
Problem: When a node is started or failed, max allowed GCP lags are
recalculated to take the change in number of data nodes into
account. However this may lead to an unintentional shutdown of a
healthy node as the threshold suddenly can decrease below the current
lag.

Fix: Delay the recalculation of new max based on number of nodes until
a succesful GCP protocol round completes.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>