WL#13505 ENABLES AUTOTEST DEFAULT CLUSTER CONFIGURATION
Modifies autotest-run.sh configuration search to include a default
configuration that can be used when running tests.
Renames vigdis05 configurations to be matched as the default
ones to use (following what we already do in practice).

This change enables running autotest-run.sh without having to manually
create a configuration file or copy an existent one to a format that
autotest-run.sh recognizes.

Change-Id: I9e59ad4ceb2bf861eea27da2d6d56983aba645da