Bug 29706669 ALTER EXCHANGE CAUSES INDEX CORRUPTION
PROBLEM

The root cause of the bug is that the order in which
the index are created in the partitioned table and
the normal table (which is to be swapped) are different.
During the alter exchange operation we basically are swapping
the se_private_data of the indexes between the partition
and the normal table,but while doing so we assume that the
order of index in both the tables are same. Since the
index order creation is different we end up swapping
se_private_data of the wrong indexes leading to corruption.

Fix

Swap the se_private_data between indexes according
to sorted index names.

RB#22947

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-by: Gopal Shankar <gopal.shankar@oracle.com>