Bug#30310265: VIEWS LOSE THE CHARACTER SET OF JSON_TABLE'S PATH ARGUMENTS
JSON_TABLE takes some string literal arguments for JSON paths and
default values. The character sets of those string literals are not
preserved when creating a view. Also, it is not possible to use a
character set introducer to provide literals in another character set
than the character set of the current session.

This patch changes the grammar rules for JSON_TABLE so that they use
text_literal instead of TEXT_STRING_sys for the string literals. This
makes the parser create Item_string objects for the literals instead
of LEX_STRING objects. The Item_string object preserves the
information about the character set of the literal.

Change-Id: I61ae0fd2bbf5d974be0f4c6a7d6e2658e7338a9a