Bug#30237657 NDB_RESTORE FAILS DUE TO EXISTING NDB_SQL_METADATA TABLE
Change ndb_restore so that the table mysql.ndb_sql_metadata by default
is not restored. Since there are times when a user would want to
restore the mysql users and grants stored in that table, also add a
new command-line option --include-stored-grants to override the default
and enable restoring its data or metadata.

This problem was previously the subject of bug#29534239. As part of that
fix, "exclude-databases=mysql" was added as an ndb_restore option in my.cnf
for MTR test suites ndb, ndb_binlog, ndb_rpl, and ndb_ddl. That option is
removed in this patch, and the tests should now succeed without it.

The ndb_sql_metadata table has an ordered index named "PRIMARY", and
ndb_restore was missing the logic to correctly handle an ordered index
on a system table, which is also added in this patch.

Change-Id: Iec11b6224a917ade9b48cc2120437dfeac57db61