Bug #30258536 DEADLOCK BETWEEN LOB::ALLOC_LOB_PAGE AND ROW_UNDO_MOD_CLUST_LOW
This bug is an analog of Bug #27777959 INNODB: ASSERTION FAILURE: UT0UT.CC DUE TO DEADLOCK AT FSP_PAGE_CREATE(),
but instead of `lob::z_rollback()`, this one involves `lob::rollback()`.
It's a regression caused by the fix to Bug #29846292 ROLLBACK OF BIG TRANSACTION DUE TO CONFLICT REMAINS IN HUNG STATE,
as now `lob::rollback()` uses internal `local_mtr` to free pages, and before it used a "one big" `mtr` provided by the caller.
This difference means that the space latch required to free a page is now acquired and released locally, and thus the
`mtr` provided by the caller does not posses the space latch. This is important, because a subsequent attempt to
latch the space, might result in a deadlock with a thread which already owns the space latch, and tries to latch the
freed page.
A very similar issue happened in the past when `z_rollback()` started to use `local_mtr` instead of `mtr`.
The solution is also similar:
We make sure that we do not latch the first page of the lob, from the main mini transaction,
before we call rollback().

However this alone is not enough. Recent fix to Bug #30178056 .IBD FILE GROWS INDEFINITELY FOR TABLES WITH BLOB COLUMNS
introduced a call to `mark_not_partially_updatable()` during rollback, which might latch the first page of the lob
in case the old lob was short enough to fit inline in the row. This leads to similar problems to the one described above
for both zLOBs and regular LOBs, as again, the first page of (z)LOB is latched before the space.
So, another strategy to avoid deadlock is employed in this patch: if we know we will need to latch a space in the future,
we latch it before manipulating the LOB - this way the order of latching is the same in the ROLLBACK as in INSERT,
and this particular deadlock scenario is impossible.

Reviewed-by: Rahul Agarkar <rahul.agarkar@oracle.com>
RB:22948