Bug#30271792 "SERVER CRASH DURING ALTER TABLER ADD COLUMN IN DIAGNOSTICS_AREA::SET_OK_STATUS".
Attempt to add column with expression default which was not valid
to an empty temporary table resulted in assertion failure.

The problem affected only debug builds.

The problem has occurred because ALTER TABLE ... ALGORITHM=COPY
code tried to store invalid expression default in record buffer
before even trying to copy rows. The error which was generated
in this case was ignored, because the rows copying code which
would have normally caught it was not called. Nor the error was
reset by binary logging code as it has happened for empty
base tables.

This patch solves the problem by avoiding storing expression
defaults into record buffer before processing any rows.
This was unnecessary work in any case as values of such defaults
need to be re-evaluate for each row since they might depend
on values in other columns in general case.

Reviewed-by: Catalin Besleaga <catalin.besleaga@oracle.com>
Reviewed-by: Sivert Sorumgaard <sivert.sorumgaard@oracle.com>