Bug #30200250 [noclose] : DOWNGRADE TESTS FAIL DUE TO BINARY CONFIGURATION
Add function to check whether the versions handled are
compatible with the binary configuration.

WL#12453 introduced a new configuration in NDB which makes the
binary configuration of cluster with version equal to or greater
than 8.0.18 incompatible with the cluster of version below 8.0.18.
This creates a requirement to restart the management nodes with
--initial option. So, downgrade tests are executed only after the
following compatibility check is made.

Downgrade from >= 8.0.18 to 8.0.18 should work without --initial
Downgrade from >= 8.0.18 to < 8.0.18 only works with --initial
Downgrade from < 8.0.18 to < 8.0.18 should work without --initial

Now, as a first step, for incompatible config versions, the test cases are
skipped.

Reviewed by: Mauritz Sundell <mauritz.sundell@oracle.com>