Bug #29520353 NDB : MODIFY SIGNALSENDER TO AVOID BUFFERING FREQUENT IRRELEVANT SIGNALS
SignalSender is used to send and received signals from C++ programs
such as NdbApi/NdbInfo and ndb_mgmd.

In most use cases, there are frequently received signals such as
SUB_GCP_COMPLETE_REP and API_REGCONF, which are not relevant to
the use case of the SignalSender.

To avoid spending cpu and memory resources on buffering these signals,
this patch implements an option to filter them out, which defaults
to on.

The ConfigManager component of ndb_mgmd needs to receive API_REGCONF
currently, so it disables the filtering.

Approved by : Martin Skold <martin.skold@oracle.com>