Bug#30326020: SUBPARTITIONING NOT REFLECTED IN SDI
The list of subpartitions in partition objects was not serialized and
hence was not part of the SDI.

This patch adds serialization and deserialization of the subpartitions,
and does some minor refactoring, e.g. changing 'sub_partitions' to
'subpartitions' in the list names to allow the json key 'subpartitions'
to conform both to the internal list name and the SQL statement keyword.

Due to the resulting change of the SDI format, the SDI version number
is incremented.

Change-Id: If9032e37d01a2269bf69867fef6e88b76b64357f