Bug#30312874: ASSERTION FAILURE AT MY_MESSAGE_SQL() IN DEBUG BUILDS
In several places in the server we tried to send a message
registered with an error-code intended for use with the
error-log (error-code >= 10,000) to the client. Our assert()
caught this. This patch creates a variant of the error message
in question in the range for messages intended for sending to
clients, and updates the various places in the code to use
this new error-code.

Approved by: Nisha Gopalakrishnan <nisha.gopalakrishnan@oracle.com>