Bug#30344462: CHANGING DEFAULT ENCRYPTION FOR A SCHEMA WILL RESET DEFAULT COLLATION
When altering only default encryption in an ALTER SCHEMA statement,
the default character set and collation for the schema would be
reset to the system default.

This patch will alter the default character set and collation for
a schema only if the corresponding clause was explicitly submitted
in the ALTER SCHEMA statement.

Additionally, if the altered schema is the currently used schema, the
state information representing the default collation will be updated
only if the relevant clause is actually present in the statement.

Change-Id: I0020bf602737914a51b9e7121e9762231720f509