Bug #30353847: CHANGE TABLE_LIST LISTS AWAY FROM LINKED LISTS
TABLE_LIST lists are used frequently in the optimizer, and are currently
chained together using a linked list, represented by the legacy List class.
However, while we frequently manipulate the front and back of these lists,
we rarely ever manipulate the middle of them (in fact, we only do so at
one specific spot, namely parenthesis removal), and since they are short
(always less than 61 elements, usually more like 3–4), that is a cheap
operation. This makes a linked list the wrong choice of data structure;
it means a lot of memory overhead (due to extra pointers) and more cache
misses (due to pointer chasing). A deque is a more suitable data structure.

Thus, change List<TABLE_LIST> to memroot_deque<TABLE_LIST *> everywhere;
memroot_deque is an std::deque with a Mem_root_allocator. This lets us forego
all the pointer chasing, and also makes for more C++-standardized interfaces
than the custom List class.

Point selects are benchmarked to +1.6–1.9% qps depending on thread count,
but this is likely to be mostly build noise. We also tried a version with
Mem_root_array<TABLE_LIST *>, but it benchmarked worse.

Change-Id: I472d6eebb540126ababe9f8be120d4628b7809ad