Bug#30344893 MTR : AVOID PROBLEMS WITH NDBD NODEID ALLOCATION
PROBLEM
=======

Having data nodes race for nodeids means that there is misalignment
between the MTR ndbd.1, ndbd.2 et al directory naming, and the
nodeids of nodes running in those directories.

This can usually be worked around by checking the output in the
ndbd.log in each directory, but this is easily overlooked and adds
unnecessary mental load.

Additionally it requires that all ndbd directories are created
prior to starting any process, as the process to directory mapping
is not deterministic.

This patch attempts to avoid this by enforcing the rule that the
MTR config name (e.g. cluster_config.ndbd.<x>.<y> implies that
the nodeid should be x.  This is used to set the ndb-nodeid
parameter to the starting ndb[mt]d process which should result in
alignment between the directory names and nodeids.

This may not work for some configurations, and may affect the
scenarios tested, wrt nodeid allocation etc.

SOLUTION
========

Per instance of ndbd/ndbmtd a ndb-nodeid entry is generated in
corresponding ndbd.<x>.<y> section in configuration, the value will match
the NodeId value set or implied in cluster_config.ndbd.<x>.<y>
configuration section.

Together with the default behaviour to imply NodeId=<x> for each
cluster_config.ndbd.<x>.<y> configuration section that will match the
number in ndbd directory with its node id.

Still, if one in some test case need to override the default assignments
of node id one can explicitly set NodeId in cluster_config.ndbd.<x>.<y>
configuration section and that will be matched with generated ndb-nodeid
setting in ndbd.<x>.<y> section.

To implement the above one also needed to change a few other things.

When MTR started a data node it passed the cluster suffix as group suffix
to ndbd/ndbmtd program instead of an instance specific group suffix (i.e.
.<y>=.1 instead of .<x>.<y>=.1.1).  That made it impossible to do instance
specific configuration of command line options (while it made it easy for
ndbd program to get the ndb-connectstring setting from mysql_cluster.<y>).

That was changed and now the group suffix passed down to ndbd is instance
specific, which also made ndbd not reading the mysql_cluster.<y> so the
ndb-connectstring must also be set in all ndbd instance specific
configuration sections ndbd.<x>.<y>.

As comparision this was already done that way for mysqld.

But for mysqld we can not do the matching of NodeId and ndb-nodeid since
mysqld can be configured to use several api node ids using the connection
pool.

Note that for ndb_mgmd one can not currently do the same change since it
uses the same suffix as group suffix for command line options as for
cluster suffix reading the cluster configuration from cluster_config
section.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>